+++ cpp/tests/modulation/sine.cpp
@@ -3,7 +3,7 @@
 // Created Date: 26/09/2023
 // Author: Shun Suzuki
 // -----
-// Last Modified: 01/12/2023
+// Last Modified: 08/12/2023
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -34,7 +34,23 @@ TEST(Modulation, Sine) {
   }
 
   ASSERT_TRUE(
-      autd.send_async(autd3::modulation::Sine(150).with_sampling_config(autd3::internal::SamplingConfiguration::from_frequency_division(512)))
-          .get());
+      autd.send_async(autd3::modulation::Sine(150).with_sampling_config(autd3::internal::SamplingConfiguration::from_frequency_division(512))).get());
   for (auto& dev : autd.geometry()) ASSERT_EQ(512, autd.link().modulation_frequency_division(dev.idx()));
+}
+
+TEST(Modulation, SineWithMode) {
+  auto autd = create_controller();
+
+  ASSERT_TRUE(autd.send_async(autd3::modulation::Sine(150).with_mode(autd3::internal::native_methods::SamplingMode::SizeOptimized)).get());
+
+  for (auto& dev : autd.geometry()) {
+    auto mod = autd.link().modulation(dev.idx());
+    std::vector<uint8_t> mod_expect{127, 156, 184, 209, 229, 244, 252, 254, 249, 237, 219, 197, 170, 142,
+                                    112, 84,  57,  35,  17,  5,   0,   2,   10,  25,  45,  70,  0};
+    ASSERT_TRUE(std::ranges::equal(mod, mod_expect));
+  }
+
+  ASSERT_THROW(autd.send_async(autd3::modulation::Sine(100.1).with_mode(autd3::internal::native_methods::SamplingMode::ExactFrequency)).get(),
+               autd3::internal::AUTDException);
+  ASSERT_TRUE(autd.send_async(autd3::modulation::Sine(100.1).with_mode(autd3::internal::native_methods::SamplingMode::SizeOptimized)).get());
 }