+++ src/autd3/src/modulation/mod.rs
@@ -4,7 +4,7 @@
  * Created Date: 31/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 22/11/2023
+ * Last Modified: 08/12/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022-2023 Shun Suzuki. All rights reserved.
@@ -14,6 +14,7 @@
 mod cache;
 mod fourier;
 mod radiation_pressure;
+mod sampling_mode;
 mod sine;
 mod square;
 mod r#static;
@@ -24,6 +25,7 @@ pub use cache::IntoCache;
 pub use fourier::Fourier;
 pub use r#static::Static;
 pub use radiation_pressure::IntoRadiationPressure;
+pub use sampling_mode::SamplingMode;
 pub use sine::Sine;
 pub use square::Square;
 pub use transform::IntoTransform;